refactor(botonic-react): create a file for opened emoji picker component
