+++ packages/botonic-react/src/webchat/webchat-input-panel/emoji-picker.tsx
@@ -57,30 +57,3 @@ export const EmojiPicker = ({
     </>
   )
 }
-
-interface OpenedEmojiPickerProps {
-  onClick: () => void
-  onEmojiClick: (event: any) => void
-}
-
-export const OpenedEmojiPicker = ({
-  onClick,
-  onEmojiClick,
-}: OpenedEmojiPickerProps) => {
-  const { ref, isComponentVisible } = useComponentVisible(true, onClick)
-  return (
-    <div ref={ref as LegacyRef<HTMLDivElement>}>
-      {isComponentVisible && (
-        <OpenedEmojiPickerContainer role={ROLES.EMOJI_PICKER}>
-          <Picker
-            width='100%'
-            height='19rem'
-            previewConfig={{ showPreview: false }}
-            onEmojiClick={onEmojiClick}
-            autoFocusSearch={false}
-          />
-        </OpenedEmojiPickerContainer>
-      )}
-    </div>
-  )
-}
