+++ packages/botonic-react/src/webchat/webchat-input-panel/index.tsx
@@ -7,7 +7,8 @@ import { WebchatContext } from '../../contexts'
 import { getFullMimeWhitelist } from '../../message-utils'
 import { DeviceAdapter } from '../devices/device-adapter'
 import { Attachment } from './attachment'
-import { EmojiPicker, OpenedEmojiPicker } from './emoji-picker'
+import { EmojiPicker } from './emoji-picker'
+import { OpenedEmojiPicker } from './opened-emoji-picker'
 import { PersistentMenu } from './persistent-menu'
 import { SendButton } from './send-button'
 import { UserInputContainer } from './styles'
